Move util function to agency utils
