+++ src/onegov/agency/collections/agencies.py
@@ -1,6 +1,5 @@
-from onegov.agency.collections.collection_utils import \
-    filter_modified_or_created
 from onegov.agency.models import ExtendedAgency
+from onegov.agency.utils import filter_modified_or_created
 from onegov.core.collection import GenericCollection, Pagination
 from onegov.people import AgencyCollection
 from sqlalchemy import or_, func
