+++ src/onegov/agency/utils.py
@@ -1,5 +1,8 @@
+import operator
 from email.headerregistry import Address
 from markupsafe import escape, Markup
+from sqlalchemy import func
+
 from onegov.core.mail import coerce_address
 from onegov.people.models import Agency, Person
 
@@ -87,3 +90,28 @@ def get_html_paragraph_with_line_breaks(text):
     return Markup('<p>{}</p>'.format(
         '<br>'.join(escape(line) for line in str(text).splitlines())
     ))
+
+
+def filter_modified_or_created(query, relate, comparison_property,
+                               collection_class):
+    ops = {
+        '>': operator.gt,
+        '<': operator.lt,
+        '>=': operator.ge,
+        '<=': operator.le,
+        '==': operator.eq,
+    }
+    assert relate in ops.keys()
+
+    return query.filter(
+        ops[relate](
+            # if 'modified' is not set comparison is done against 'created'
+            func.coalesce(
+                func.date_trunc('minute',
+                                collection_class.modified),
+                func.date_trunc('minute',
+                                collection_class.created),
+            ),
+            comparison_property
+        )
+    )
